Unify RekorPubKeys and CTLogPubKeys, refactor.
