+++ pkg/cosign/fulcioverifier/ctutil/ctutil_test.go
@@ -15,20 +15,13 @@
 package ctutil
 
 import (
-	"crypto/x509"
 	"encoding/base64"
-	"encoding/pem"
 	"testing"
 
 	ct "github.com/google/certificate-transparency-go"
 	"github.com/google/certificate-transparency-go/testdata"
 	"github.com/google/certificate-transparency-go/tls"
-	ttestdata "github.com/google/certificate-transparency-go/trillian/testdata"
 	"github.com/google/certificate-transparency-go/x509util"
-)
-
-var (
-	demoLogID = [32]byte{19, 56, 222, 93, 229, 36, 102, 128, 227, 214, 3, 121, 93, 175, 126, 236, 97, 217, 34, 32, 40, 233, 98, 27, 46, 179, 164, 251, 84, 10, 60, 57}
 )
 
 func TestLeafHash(t *testing.T) {
@@ -364,22 +357,5 @@ func TestContainsSCT(t *testing.T) {
 				t.Errorf("ContainsSCT(_,_) = %t, nil, want %t, nil", got, test.want)
 			}
 		})
-	}
-}
-
-func TestGetCTLogID(t *testing.T) {
-	block, _ := pem.Decode([]byte(ttestdata.DemoPublicKey))
-	pk, err := x509.ParsePKIXPublicKey(block.Bytes)
-	if err != nil {
-		t.Fatalf("unexpected error loading public key: %v", err)
-	}
-
-	got, err := GetCTLogID(pk)
-	if err != nil {
-		t.Fatalf("error getting logid: %v", err)
-	}
-
-	if want := demoLogID; got != want {
-		t.Errorf("logID: \n%v want \n%v", got, want)
 	}
 }
